Complete overhaul of the project
I've changed a lot of the formatting to be in line with common Python practices (getting rid of camel case for variables, changing function names, etc.) In addition, I've added a function that can check the orbit of a function of a matrix up to a number k. This is found by calculating f^k(x) at the beginning (remember, this is fairly efficient due to the square-and-multiply method), and then checking, for all m in (1, k), if f^m(x) == f^k(x).